Changed name to Select and removed props
