+++ src/index.ts
@@ -1,6 +1,6 @@
 import Button from "./components/Button";
 import TextInput from "./components/TextInput";
 import ToggleButton from "./components/ToggleButton"
-import SelectInput from "./components/SelectInput"
+import SelectInput from "./components/Select"
 
 export { Button, TextInput, ToggleButton, SelectInput };
